requirements.txt
absl-py
annotated-types
anyio
argon2-cffi
argon2-cffi-bindings
arrow
asttokens
astunparse
async-lru
attrs
babel
beautifulsoup4
black
bleach
blis
catalogue
certifi
cffi
chardet
charset-normalizer
click
cloudpathlib
colorama
comm
confection
contourpy
cramjam
cycler
cymem
debugpy
decorator
defusedxml
dotenv
duckdb
emoji
et_xmlfile
executing
fastjsonschema
fastparquet
filelock
flatbuffers
fonttools
fqdn
fsspec
gast
gitdb
GitPython
google-pasta
grpcio
h11
h5py
httpcore
httpx
huggingface-hub
idna
imageio
ipykernel
ipython
ipython_pygments_lexers
ipywidgets
isoduration
isort
jedi
Jinja2
joblib
json5
jsonpointer
jsonschema
jsonschema-specifications
jupyter
jupyter-console
jupyter-events
jupyter-lsp
jupyter-server-mathjax
jupyter_client
jupyter_core
jupyter_server
jupyter_server_terminals
jupyterlab
jupyterlab_pygments
jupyterlab_server
jupyterlab_widgets
keras
kiwisolver
langcodes
language_data
lark
libclang
lightgbm
marisa-trie
Markdown
markdown-it-py
MarkupSafe
matplotlib
matplotlib-inline
mdurl
mistune
ml_dtypes
mpmath
murmurhash
mypy_extensions
namex
narwhals
nbclient
nbconvert
nbdime
nbformat
nest-asyncio
networkx
nltk
notebook
notebook_shim
numpy
openpyxl
opt_einsum
optree
outcome
overrides
packaging
pandas
pandocfilters
parso
pathspec
pillow
platformdirs
plotly
preshed
prometheus_client
prompt_toolkit
protobuf
psutil
pure_eval
pyarrow
pycparser
pydantic
pydantic_core
Pygments
pyparsing
pyreadstat
PySocks
python-dateutil
python-dotenv
python-json-logger
pytokens
pytz
pywin32
pywinpty
PyYAML
pyzmq
referencing
regex
requests
rfc3339-validator
rfc3986-validator
rfc3987-syntax
rich
rpds-py
safetensors
scikit-learn
scipy
seaborn
selenium
Send2Trash
sentence-transformers
setuptools
shellingham
six
smart_open
smmap
sniffio
sortedcontainers
soupsieve
spacy
spacy-legacy
spacy-loggers
srsly
stack-data
sympy
tensorboard
tensorboard-data-server
tensorflow
termcolor
terminado
thinc
threadpoolctl
tinycss2
tokenizers
torch
tornado
tqdm
traitlets
transformers
trio
trio-websocket
typer
types-python-dateutil
typing-inspection
typing_extensions
tzdata
uri-template
urllib3
wasabi
wcwidth
weasel
webcolors
webencodings
websocket-client
Werkzeug
wheel
widgetsnbextension
wrapt
wsproto
xgboost
XlsxWriter
